+++ CHANGELOG.md
@@ -15,6 +15,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Realm attributes in configuration file overwrite the realm's state when the realm is updated.
+- Custom realm attributes not updatable.
 
 ## [4.0.1] - 2021-06-19